use bash instead of sh, to use "type -P' instead of "which"
"command -v" would not be sufficient. The behavior of which varies wildly; for
example, GNU which prints a message when something isn't found.

The old version of "that" assumed the version of which provided by debianutils.

I believe there is no reasonable portable way to achieve what I was trying to
achieve with which, but type -P does it as well as anything I can think of.

So I'm making it a bash script.

This change also facilitates a fix for the code smell of passing the result of
an expansion as the first non-option argument to printf, which I could still
have fixed in pure sh... but less nicely.