added a script to run shellcheck on the scripts here (including itself)
This is taken from the setup-cgit repository, and improved.

I plan to replace that script with this one soon, after which both repositories
will have identical "check" scripts. Perhaps this should go in a repository.
Perhaps it should be one of the scripts *this* repository installs.

Is "check" an "inspector" for jacques-clouseau purposes? I'm not sure.